Update PotentialDatasets
Hello there!
As you may have notice, I made slight changes to the file. I rearranged the data sets in an alphabetical style to keep this neat and tidy. I also added subcategories to the subject categorizes. For example, science is a category and biology is a subcategory in science and plants is a sub-subcategory in biology which is a subcategory in science. This structures the file like a tree diagram that branches out into different, yet, similar and related subjects.